feat: update postgres driver to use proto types and updated README.md
