+++ README.md
@@ -12,6 +12,8 @@
 - [3. Data Sources](#3-data-sources)
   - [3.1. YAML](#31-yaml)
   - [3.2. Postgres](#32-postgres)
+    - [3.2.1. Grove Portal DB Compatibility](#321-grove-portal-db-compatibility)
+    - [3.2.2. SQLC Autogeneration](#322-sqlc-autogeneration)
 - [4. gRPC Proto File](#4-grpc-proto-file)
 
 ## 1. Introduction
@@ -105,7 +107,32 @@ endpoints:
 
 If the `POSTGRES_CONNECTION_STRING` environment variable is set, PADS will connect to the specified Postgres database.
 
-The connected Postgres database must contain the tables and schema defined in the [`postgres/sqlc/schema.sql`](postgres/sqlc/schema.sql) file.
+#### 3.2.1. Grove Portal DB Compatibility
+
+The [Postgres Driver schema file](postgres/sqlc/schema.sql) uses tables from the existing Grove Portal database schema, allowing PATH to source its authorization data from the existing Grove Portal DB. 
+
+It converts the data stored in the `portal_applications` table and its associated tables into the `proto.GatewayEndpoint` format expected by PATH's Go External Authorization Server.
+
+It also listens for updates to the Grove Portal DB and streams updates to the `Go External Authorization Server` in real time as changes are made to the connected Postgres database.
+
+[For the full Grove Portal DB schema, refer to the database schema defined in the Portal HTTP DB (PHD) repository](https://github.com/pokt-foundation/portal-http-**db**/blob/master/postgres-driver/sqlc/schema.sql).
+
+#### 3.2.2. SQLC Autogeneration
+
+<div align="center">
+<a href="https://docs.sqlc.dev/en/stable">
+<img src="https://docs.sqlc.dev/en/stable/_static/logo.png" alt="SQLC logo" width="150"/>
+<div>https://docs.sqlc.dev/en/stable</div>
+</a>
+</div>
+<br/>
+
+The Postgres Driver uses `SQLC` to autogenerate the Postgres Go code needed to interact with the Postgres database
+
+
+The Make target `make gen_sqlc` will regenerate the Go code from the SQLC schema file. This will output autogenerated code to the [postgres/sqlc/schema.sql](postgres/sqlc/schema.sql) and [postgres/sqlc/query.sql](postgres/sqlc/query.sql) files to the [postgres/sqlc](postgres/sqlc) directory.
+
+`SQLC` configuration is defined in the [postgres/sqlc/sqlc.yaml](postgres/sqlc/sqlc.yaml) file.
 
 ## 4. gRPC Proto File
